// src/app/data-grid/models/grid-column.ts
export interface GridColumn {
  name : string;
  caption: string;
  value?: any;
  width?: number;
  visible: boolean;
}

export class ColumnBuilder {
  // static build() : GridColumn {
  //   return {
  //     name : 'Employee'
  // value?: any;
  // width?: number;
  // visible: boolean;
  //   };
  // }
}